Implemented Member Initialization Lists
Implemented member initialization lists on my two constructors. This was done simply to use a new tool I cam across but also to familiarize myself with the practice of using them for later classes. I understand they can be advantageous/necessary when you are dealing with constant member variables and  in cases of composition (using an instance one class object inside of an object of a different class).